Separate production and integration GCP access
At the moment the "govuk-gcp-access" google groups grants admin access
to all three search api v2 gcp environments. This isn't ideal as on
GOV.UK we have rules about when someone should be granted
[production access].

The idea is to create two new google groups:

* govuk-gcp-access-production
* govuk-gcp-access-integration

govuk-gcp-access-production will be a copy of govuk-gcp-access, with
members copied across. It will grant members admin access to the
production and staging search api V2 projects on GCP.

govuk-gcp-access-integration will grant members admin access to the
integration search api V2 project on GCP, but not staging and production.
govuk-gcp-access-production will be a member of
govuk-gcp-access-integration so that members also have access to
integration.

I'm not sure how the resources created here are associated with their
respective modules (environment_production, environment_staging, environment_integration) in  "./modules/search-api-v2" and if anything
else needs to be amended.

I'm not sure if this change is correct and will do what I expect.

If it is, there are still things to consider:

Is this the best way to manage access?
Should staging have its own group to manage access?
Is there a way to dry this up?

[production access]: https://docs.publishing.service.gov.uk/manual/rules-for-getting-production-access.html#when-you-get-production-admin-access