Search API V2: add evaluations dataset and table
For storing the results of evaluations that are created using the
built-in Vertex AI Search service. These will replace the ones that we
computed elsewhere.

This commit defines a dataset, a table, and a schema for the table.

The table is an external table, which means that no data is stored
natively in BigQuery. Queries will fetch data from JSON files in the
bucket that was created by #2485.

The names and schemas of those JSON files must conform to the BigQuery
table schema that is in this commit, and the `source_uri_prefix` that is
configured in terraform.

The `source_uri_prefix` causes BigQuery to interpret directory names as
column values. For example, the file
`search-api-v2-integration_vais_evaluation_output/judgement_list=explicit/partition_date=2025-08-01/create_time=2025-08-28T10:53:20/evaluation_id=0403d479-10e5-4724-9ca1-873f4a5b4008/results.json`
will be imported with the date value `2025-08-01` in column called
`partition_date`.