Search API V2: move evaluation id/name into JSON
1. Remove the `evaluation_id` (a UUID) from the name of the bucket
   object, where it is currently used as a hive partition
2. Put the full `evaluation_name` into the JSON of the object itself.

Advantages: It will be easier to join the table to a separate table of
evaluations via the full name than via the UUID alone. The full
`evaluation_name` contains `/`, which would be difficult to use as part
of an object name.

Disadvantages: Less efficient queries, because of not partitioning by
evaluation id/name. Less easy to find the bucket object that relates to
a particular evaluation (though it can still be done because the
`create_time` will be unique to an evaluation).

3. Remove the hive partition columns from the table schema. Fixes error
   during `terraform apply`:

```
Error: googleapi: Error 400: Field, judgement_list, is present in both Table Schema and Hive-Partition key., invalid
```