Optionally generate release notes
This adds an input to the Release workflow to automatically generate
release notes (see https://docs.github.com/en/repositories/releasing-projects-on-github/automatically-generated-release-notes).

While we practice continuous delivery, so one PR usually results in one
release, this could still be useful to link releases to PRs. For now,
I’ve made it optional with a default of true, but I’m happy to set it
to `true` for our project and see how we go.