DO NOT MERGE: create tfvars files for each root being applied
In Terraform Cloud, variablue values are managed with variable sets which are
set globally and referenced by a workspace.

Regular Terraform has no such concept. Instead we create tfvars files with
variable values in them, separated by whether the value is applicable to all
instaces of a type of environment (e.g. all development environments), or one
specific instance (e.g. the dev01 development environment)

When we apply the Terraform, we'll check for the existence of relevant tfvars
files and include them.

The values in this commit are taken from the Terraform Cloud variable set
values. I've done my best to only include those that are necessary in each
root, but I've probably added some that aren't or missed some that are.