Allow merge commits by default on new repos
Allow merge commits (Add all commits from the head branch to the base branch
with a merge commit.) is enabled for all GOV.UK repos, which is what we want,
despite  this config stating otherwise.

When a new repo is created this value is set to false which engineers found
confusing as there was no option to "Merge" the PR  but to "Squash and merge".
They had to change the Pull Request setting in the UI. Subsequent terraform
apply runs don't appear to be changing this setting so it maybe something to
look into later.

Tested by running TF plan on a new repo and this value is set correctly for a
new repo:
  # github_repository.govuk_repos["new-repo"] will be created
  + resource "github_repository" "govuk_repos" {
      + allow_auto_merge            = false
      + allow_merge_commit          = true
      + allow_rebase_merge          = true
      + allow_squash_merge          = true
      ...