Add permission to delete storage objects
A scheduled task failed to upload Search API v2 evaluation results because a bucket object with the same name already existed, and the task lacked permission to delete it before replacing it.

We would probably always want to replace an existing file anyway, so this commit grants the `storage.objects.delete` permission.

Note that permissions are granted at the project level, whereas granting them at the resource level would better meet the principle of least privilege. We intend to refactor this when we remove the redundant infrastructure of manual evaluations.