Change demote value of 5 HMRC pages.
These pages are still being surfaced and so require a stronger demote
value (-0.5 instead of -0.25)

Demote another HMRC specialist manual section. This manual section is
getting bad feedback because it is being surfaced but is generally not what
users are looking for. This means we need to demote it.

This change means we currently don't demote pages by -0.25. This PR does
not delete that code because we would like to keep the option to quickly
demote pages by that value.