Archive govuk-synthetic-test-runner
This was created as it was initially thought that the ECR pull through was not working for govuk-synthetic-test-app but this was because there wasn't a demand for the latest image tag so the image wasn't pulled through.

As the ECR pull through is now working and that repo has more historical context this repo will be retired.