Setup a barebones GCP project for govgraph
Currently the terraform for govgraph sits entierly within it's own repo (alphagov/govuk-knowledge-graph-gcp)
We'd like to move this over to terraform cloud so that our tf setup is more
in line with the rest of GOV.UK.

This borrows heavily from the GCP setup for search API V2. I think I've removed
everything that's search specific though.