Add attribution token to search event schema
The Vertex product manager has asked the search
team to include attribution tokens in the search
user events that are sent to Vertex, to enable
them to analyse relevance quality and diagnose
ranking issues more effectively. This change to
add the attribution token to the search event
table schemas in BigQuery is a prerequisite to
taking the data from GA4 and sending it to Vertex.