Disable the blue Elasticsearch cluster in Integration
A new green cluster has been instantiated in Integration and now serves
all traffic. The blue cluster is no longer required and can be safely
disabled until it is needed again for future upgrades.

This change introduces a toggle variable to enable or disable the blue
Elasticsearch cluster, and sets it to disabled for the Integration
environment only.