Mirror sync: allow s3:*Object via Athena
When a user makes a query with Amazon Athena, it reuses the credentials to act
on the user's behalf when operating on AWS S3. This means the credentials need
permission to do all sorts with objects within the bucket.

To limit the scope of that permission, we make it conditional on the requests
to S3 coming via Amazon Athena. There is an exception for s3:GetObject, which
the mirror role requires to be able to do directly.