Introduce very basic Vale prose linting setup
Vale[1] is a linting tool for the written word. It can enforce a style guide
and properties of prose in the same way a linter does for a programming
language.

This initial commit of it is deliberately dumb, and does not include a
vocabulary or style guide that would be applicable to our writing. That will
follow in subsequent commits.

The "--no-exit" flag is included for now because we don't want running it and
seeing errors to be considered a failure. We will fix the errors it finds in
time.

[1] https://vale.sh/