Prose linting: swap to RedHat style as the base style
The Google style is maintained by the creators of Vale, and bug fixes are not
frequently merged. It contains small annoyances like saying "GCP" should become
"Google Cloud Platform", but then also saying "Cloud" in "Google Cloud
Platform" should be replaced with "Google Cloud Platform".

The RedHat style is maintained by RedHad themselves, and is in active use. It
ought to be a better bet to serve as an underlying style that we further
customise.