Prose linting: allow Makefile target to monitor files for changes
Beware: Makefile gore.

To prevent recursive calls to the Make file, and remove the duplication of the
Vale command, we define a Make function valled "fn_vale", and use the "call"
builtin to call it when needed.

The target starts by running the function, and if WATCH_DOCS=true it continues
to start a file watcher using fswatch, which in turns invokes "fn_vale" on each
file change.